pyproject.toml
[project]
name = "license-plates"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "argparse>=1.4.0",
    "opencv-contrib-python>=4.11.0.86",
    "pillow>=11.2.1",
    "pytesseract>=0.3.13",
    "torch>=2.7.1",
    "ultralytics>=8.3.159",
]
